Built little object library
The previous commit did not save the changes made to the project properties. Some new changes appeared upon closing the project, so I figure they might've been saved in those.

In this commit I also attempt to make an acceleration calculation, trying to imitate drag with a simplified equation as well.

So far creating a new Object, and passing a shape through the constructor, does not seem to work. It might have something to do with the fact that I'm creating a shallow copy. Either that, or something with the inheritance I'm using.