Fixed the inheritance issue
The sprite within the object class, was a pass by value copy I believe. This means that the variable might've gotten treated as it's base class, a Shape, instead of a Square.

This has been fixed by makign it a pointer instead, in which case data such as that gets preserverd.