Added Vector2 class
Currently the Vector2 class does not seems to work, as Vector2 classes seem to end up as (0,0) if passed through enough parameters.

A new copy of a Vector2 class gets created when passed through a paremeter, so maybe it's missing a copy constructor.

Alternatively, I could make the Vector2 parameters a reference, instead of a copy. But I feel like that might not be necessary. I'll have to test out which option has better performance.