Expanded 'Shape' functionality
I added a 'Bounds' class to ease the calculation of a Shapes bounds. This will help with collision.

I also implemented an 'EntityManager' wich manager the player, and an enemy for now. This however, caused a mysterious null error which seemingly shouldn't appear, as the shape gets allocated before the draw fucntion gets called.

More research is required. :(