Fixed collision and the inexplainable error
I made a tiny mistake in the box collision detection. Every condition should be met, and not at least half of them.

On top of that, the inexplainable error was caused because the player was assigned in the EntityManager's constructor. However, because it was created in the stack, the player would most likely get deleted as soon as it leaves the constructor's scope.

Making the player a pointer fixed this, as the memory probably stays in the heap, and won't disappear at the end of the scope. If my theory is true at least.

This would also explain the error not occuring if the sprite doesn't get deleted in the Object's destructor. The sprite would still exist, so the draw function (where the error occured) could still be called.