Created EnemyManager
a new inexplainable error has occured. I'm starting to believe the error does not come to fruition due to the created object falling out of scope. I don't think it works that way. Rather, I think the case is that I'm treating the constructor of many classes as an assignment operators. I should probably convert quite a bit of constructors to assignment operators, so that I can initialise them after they are already created.